requirements.txt
numpy>=1.21
scipy>=1.7
librosa>=0.9.2
soundfile>=0.10.3.post1
openai-whisper>=20230314
torch